chore(restructure/git): Add .gitattributes etc. from skeleton-module
Removed useless directories
Made crossplatform with gitattributes and dos2unix
Added new CMakeFile
.editorconfig
.gitignore
