--- module/TemplateNPC/CMakeLists.txt
@@ -1,4 +1,0 @@
-CU_SET_PATH("CMAKE_TEMPLATE_DIR" "${CMAKE_CURRENT_LIST_DIR}")
-CU_SET_PATH("CMAKE_TEMPLATE_SRC_DIR" "${CMAKE_CURRENT_LIST_DIR}/src/")
-
-include("${CMAKE_TEMPLATE_SRC_DIR}/CMakeLists.txt")
