--- module/TemplateNPC/conf/conf.sh.dist
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-
-#
-# CUSTOM
-#
-
-DB_WORLD_CUSTOM_PATHS+=(
-	$TEMPLATENPC_ROOT"/data/sql/db-characters/"
-)
