--- module/TemplateNPC/src/template_scripts_loader.h
@@ -1,11 +1,0 @@
-#ifndef TEMPLATE_SCRIPTS_LOADER_H
-#define TEMPLATE_SCRIPTS_LOADER_H
-
-void AddSC_TemplateNPC();
-
-void AddTemplateScripts()
-{
-    AddSC_TemplateNPC();
-}
-
-#endif /* TEMPLATE_SCRIPTS_LOADER_H */
