+++ setup_git_commit_template.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+## Set a local git commit template
+git config --local commit.template ".git_commit_template.txt" ;
