+++ sql/characters/characters.sql
