+++ src/CMakeLists.txt
@@ -1,0 +1,6 @@
+CU_SET_PATH("CMAKE_TEMPLATE_SRC_DIR" "${CMAKE_CURRENT_LIST_DIR}")
+
+AC_ADD_SCRIPT("${CMAKE_TEMPLATE_SRC_DIR}/TemplateNPC.cpp")
+AC_ADD_SCRIPT("${CMAKE_TEMPLATE_SRC_DIR}/TemplateNPC.h")
+
+AC_ADD_SCRIPT_LOADER("Template" "${CMAKE_TEMPLATE_SRC_DIR}/template_scripts_loader.h")
