+++ src/TemplateNPC.h
